Consolidate project path normalization helpers
