+++ src/plugin/src/identifier-case/local-plan.js
@@ -1,8 +1,6 @@
 import { formatIdentifierCase } from "./identifier-case-utils.js";
 import { buildRenameKey } from "./plan-state.js";
 import { asArray, isNonEmptyArray } from "../../../shared/array-utils.js";
-import { toPosixPath } from "../../../shared/path-utils.js";
-import { resolveProjectPathInfo } from "../project-index/path-info.js";
 import { createMetricsTracker } from "../reporting/metrics.js";
 import {
     isNonEmptyString,
@@ -34,6 +32,7 @@ import {
 } from "./common.js";
 import { planAssetRenames, applyAssetRenames } from "./asset-renames.js";
 import { getIterableSize } from "../../../shared/utils/capability-probes.js";
+import { resolveProjectRelativeFilePath } from "../project-index/path-normalization.js";
 import { getDefaultIdentifierCaseFsFacade } from "./fs-facade.js";
 import { evaluateIdentifierCaseAssetRenamePolicy } from "./asset-rename-policy.js";
 
@@ -48,19 +47,6 @@ function getScopeDisplayName(scopeRecord, fallback = "<unknown>") {
         scopeRecord.id ??
         fallback
     );
-}
-
-function resolveRelativeFilePath(projectRoot, absoluteFilePath) {
-    const info = resolveProjectPathInfo(absoluteFilePath, projectRoot);
-    if (!info) {
-        return null;
-    }
-
-    if (!info.hasProjectRoot) {
-        return toPosixPath(info.absolutePath);
-    }
-
-    return toPosixPath(info.relativePath);
 }
 
 function createScopeGroupingKey(scopeId, fallback) {
@@ -865,7 +851,7 @@ export async function prepareIdentifierCasePlan(options) {
     let fileRecord = null;
     let relativeFilePath = null;
     if (hasLocalSupport) {
-        relativeFilePath = resolveRelativeFilePath(
+        relativeFilePath = resolveProjectRelativeFilePath(
             projectIndex.projectRoot,
             options.filepath ?? null
         );
